좋아요 / 팔로우 취소시 알림 삭제

+++ src/main/java/com/example/lit/domain/dao/user/AlertDAO.java
@@ -17,9 +17,9 @@ public class AlertDAO {
 
     public List<AlertDTO> getList(Long userNumber){ return alertMapper.alertList(userNumber); }
 
-    public Long getAlertNumber(AlertVO alertVO) {return alertMapper.getAlertNumber(alertVO); }
+    public int getAlertNumber(AlertVO alertVO) {return alertMapper.getAlertNumber(alertVO); }
 
-    public void remove(Long alertNumber) { alertMapper.remove(alertNumber);}
+    public void remove(int alertNumber) { alertMapper.remove(alertNumber);}
 
 
 }
